LPD-72635 Move common code to a Util class
