+++ modules/apps/export-import/export-import-test/src/testIntegration/java/com/liferay/exportimport/internal/content/processor/test/DefaultExportImportContentProcessorTest.java
@@ -13,6 +13,7 @@ import com.liferay.document.library.kernel.service.DLAppHelperLocalServiceUtil;
 import com.liferay.document.library.kernel.service.DLAppLocalServiceUtil;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessorRegistryUtil;
+import com.liferay.exportimport.internal.content.processor.test.util.ExportImportContentProcessorTestUtil;
 import com.liferay.exportimport.kernel.configuration.ExportImportConfigurationSettingsMapFactoryUtil;
 import com.liferay.exportimport.kernel.configuration.constants.ExportImportConfigurationConstants;
 import com.liferay.exportimport.kernel.exception.ExportImportContentValidationException;
@@ -29,23 +30,18 @@ import com.liferay.journal.model.JournalArticle;
 import com.liferay.journal.test.util.JournalTestUtil;
 import com.liferay.layout.test.util.LayoutFriendlyURLRandomizerBumper;
 import com.liferay.layout.test.util.LayoutTestUtil;
-import com.liferay.petra.function.transform.TransformUtil;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.NoSuchLayoutException;
-import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.GroupConstants;
 import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.model.LayoutSet;
 import com.liferay.portal.kernel.model.StagedModel;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.model.VirtualLayoutConstants;
-import com.liferay.portal.kernel.portlet.constants.FriendlyURLResolverConstants;
 import com.liferay.portal.kernel.repository.capabilities.ThumbnailCapability;
 import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.service.CompanyLocalServiceUtil;
 import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
 import com.liferay.portal.kernel.service.LayoutLocalServiceUtil;
 import com.liferay.portal.kernel.test.ReflectionTestUtil;
@@ -63,12 +59,12 @@ import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.FriendlyURLNormalizerUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.kernel.util.ObjectValuePair;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.PropsValues;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Time;
-import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.xml.Document;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.kernel.xml.SAXReaderUtil;
@@ -78,19 +74,14 @@ import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 import com.liferay.portal.util.PortalImpl;
 
 import java.io.File;
-import java.io.InputStream;
 import java.io.Serializable;
 
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.NavigableMap;
-import java.util.Scanner;
 import java.util.UUID;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.junit.Assert;
@@ -117,7 +108,6 @@ public class DefaultExportImportContentProcessorTest {
 	@Before
 	public void setUp() throws Exception {
 		_defaultLocale = LocaleUtil.getDefault();
-		_nondefaultLocale = _getNondefaultLocale();
 
 		UserTestUtil.setUser(TestPropsValues.getUser());
 
@@ -224,13 +214,20 @@ public class DefaultExportImportContentProcessorTest {
 	public void testExportDLReferences() throws Exception {
 		_portletDataContextExport.setZipWriter(new TestReaderWriter());
 
-		String content = _replaceParameters(
-			_getContent("dl_references.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "dl_references.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		_exportImportContentProcessor.validateContentReferences(
 			_stagingGroup.getGroupId(), content);
 
-		List<String> urls = _getURLs(content);
+		List<String> urls = ExportImportContentProcessorTestUtil.getURLs(
+			content);
 
 		content = _exportImportContentProcessor.replaceExportContentReferences(
 			_portletDataContextExport, _referrerStagedModel, content, true,
@@ -284,13 +281,20 @@ public class DefaultExportImportContentProcessorTest {
 			ServiceContextTestUtil.getServiceContext(
 				_stagingGroup.getGroupId(), TestPropsValues.getUserId()));
 
-		String content = _replaceParameters(
-			_getContent("dl_references_file_friendly_urls.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "dl_references_file_friendly_urls.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		_exportImportContentProcessor.validateContentReferences(
 			_stagingGroup.getGroupId(), content);
 
-		List<String> urls = _getURLs(content);
+		List<String> urls = ExportImportContentProcessorTestUtil.getURLs(
+			content);
 
 		content = _exportImportContentProcessor.replaceExportContentReferences(
 			_portletDataContextExport, _referrerStagedModel, content, true,
@@ -343,10 +347,17 @@ public class DefaultExportImportContentProcessorTest {
 			ServiceContextTestUtil.getServiceContext(
 				_stagingGroup.getGroupId(), TestPropsValues.getUserId()));
 
-		String content = _replaceParameters(
-			_getContent("invalid_dl_references.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "invalid_dl_references.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
-		List<String> urls = _getURLs(content);
+		List<String> urls = ExportImportContentProcessorTestUtil.getURLs(
+			content);
 
 		content = _exportImportContentProcessor.replaceExportContentReferences(
 			_portletDataContextExport, _referrerStagedModel, content, true,
@@ -389,8 +400,14 @@ public class DefaultExportImportContentProcessorTest {
 			_layoutReferencesExportImportContentProcessor,
 			"_PRIVATE_USER_SERVLET_MAPPING", "/en/");
 
-		String content = _replaceParameters(
-			_getContent("layout_references.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "layout_references.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		_exportImportContentProcessor.validateContentReferences(
 			_stagingGroup.getGroupId(), content);
@@ -460,8 +477,14 @@ public class DefaultExportImportContentProcessorTest {
 			_layoutReferencesExportImportContentProcessor,
 			"_PRIVATE_USER_SERVLET_MAPPING", "/en/");
 
-		String content = _replaceParameters(
-			_getContent("layout_references.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "layout_references.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		_exportImportContentProcessor.validateContentReferences(
 			_stagingGroup.getGroupId(), content);
@@ -512,8 +535,8 @@ public class DefaultExportImportContentProcessorTest {
 	@Test
 	public void testExportLinksToLayouts() throws Exception {
 		String content = _replaceLinksToLayoutsParameters(
-			_getContent("layout_links.txt"), _stagingPrivateLayout,
-			_stagingPublicLayout);
+			ExportImportContentProcessorTestUtil.getContent("layout_links.txt"),
+			_stagingPrivateLayout, _stagingPublicLayout);
 
 		content = _exportImportContentProcessor.replaceExportContentReferences(
 			_portletDataContextExport, _referrerStagedModel, content, true,
@@ -531,7 +554,8 @@ public class DefaultExportImportContentProcessorTest {
 	public void testExportLinksToURLsWithStopCharacters() throws Exception {
 		String path = RandomTestUtil.randomString();
 
-		String content = _getContent("url_links.txt");
+		String content = ExportImportContentProcessorTestUtil.getContent(
+			"url_links.txt");
 
 		content = content.replaceAll("PATH", path);
 
@@ -568,8 +592,9 @@ public class DefaultExportImportContentProcessorTest {
 			RandomTestUtil.randomString());
 
 		String content = _replaceLinksToLayoutsParameters(
-			_getContent("layout_links_user_group.txt"), privateLayout,
-			publicLayout);
+			ExportImportContentProcessorTestUtil.getContent(
+				"layout_links_user_group.txt"),
+			privateLayout, publicLayout);
 
 		content = _exportImportContentProcessor.replaceExportContentReferences(
 			portletDataContextExport, journalArticle, content, true, true);
@@ -599,8 +624,14 @@ public class DefaultExportImportContentProcessorTest {
 
 		_portletDataContextExport.setZipWriter(new TestReaderWriter());
 
-		String content = _replaceParameters(
-			_getContent("dl_references.txt"), fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			fileEntry, "dl_references.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		_exportImportContentProcessor.validateContentReferences(
 			_stagingGroup.getGroupId(), content);
@@ -662,8 +693,14 @@ public class DefaultExportImportContentProcessorTest {
 	public void testImportLayoutReferencesOnExternalGroupWithDifferentUUID()
 		throws Exception {
 
-		String content = _replaceParameters(
-			_getContent("layout_references.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "layout_references.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		_exportImportContentProcessor.validateContentReferences(
 			_stagingGroup.getGroupId(), content);
@@ -719,12 +756,12 @@ public class DefaultExportImportContentProcessorTest {
 	@Test
 	public void testImportLinksToLayouts() throws Exception {
 		String content = _replaceLinksToLayoutsParameters(
-			_getContent("layout_links.txt"), _stagingPrivateLayout,
-			_stagingPublicLayout);
+			ExportImportContentProcessorTestUtil.getContent("layout_links.txt"),
+			_stagingPrivateLayout, _stagingPublicLayout);
 
 		String liveContent = _replaceLinksToLayoutsParameters(
-			_getContent("layout_links.txt"), _livePrivateLayout,
-			_livePublicLayout);
+			ExportImportContentProcessorTestUtil.getContent("layout_links.txt"),
+			_livePrivateLayout, _livePublicLayout);
 
 		content = _exportImportContentProcessor.replaceExportContentReferences(
 			_portletDataContextExport, _referrerStagedModel, content, true,
@@ -761,7 +798,8 @@ public class DefaultExportImportContentProcessorTest {
 		layoutPlids.put(
 			_stagingPublicLayout.getPlid(), importedPublicLayout.getPlid());
 
-		String content = _getContent("layout_links_ids_replacement.txt");
+		String content = ExportImportContentProcessorTestUtil.getContent(
+			"layout_links_ids_replacement.txt");
 
 		String expectedContent = _replaceLinksToLayoutsParameters(
 			content, importedPrivateLayout, importedPublicLayout);
@@ -798,7 +836,7 @@ public class DefaultExportImportContentProcessorTest {
 
 		layouts.put(3L, importedPrivateLayout);
 
-		String contentInFile = _getContent(
+		String contentInFile = ExportImportContentProcessorTestUtil.getContent(
 			"layout_links_in_layoutset_prototype.txt");
 
 		String content = _replaceLinksToLayoutsParametersInLayoutSetPrototype(
@@ -830,8 +868,14 @@ public class DefaultExportImportContentProcessorTest {
 
 		portalUtil.setPortal(portalImpl);
 
-		String content = _replaceParameters(
-			_getContent("invalid_layout_references.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "invalid_layout_references.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		String[] layoutReferences = StringUtil.split(
 			content, StringPool.NEW_LINE);
@@ -874,8 +918,14 @@ public class DefaultExportImportContentProcessorTest {
 	public void testReplaceExportContentReferencesWithFileEntryInTrash()
 		throws Exception {
 
-		String content = _replaceParameters(
-			_getContent("journal-content.xml"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "journal-content.xml", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		FileEntry deletedFileEntry = DLAppLocalServiceUtil.addFileEntry(
 			null, TestPropsValues.getUserId(), _stagingGroup.getGroupId(),
@@ -927,7 +977,12 @@ public class DefaultExportImportContentProcessorTest {
 		Map<Locale, String> nameMap = new HashMap<>();
 		Map<Locale, String> friendlyURLMap = new HashMap<>();
 
-		for (Locale locale : new Locale[] {_defaultLocale, _nondefaultLocale}) {
+		for (Locale locale :
+				new Locale[] {
+					_defaultLocale,
+					ExportImportContentProcessorTestUtil.getNondefaultLocale()
+				}) {
+
 			String name = RandomTestUtil.randomString(
 				LayoutFriendlyURLRandomizerBumper.INSTANCE,
 				NumericStringRandomizerBumper.INSTANCE,
@@ -1027,28 +1082,6 @@ public class DefaultExportImportContentProcessorTest {
 		Assert.assertTrue(content, content.contains(sb.toString()));
 	}
 
-	private String _duplicateLinesWithParamNames(
-		String content, String[] findParams, String[] addParams) {
-
-		if (StringUtil.indexOfAny(content, findParams) <= -1) {
-			return content;
-		}
-
-		List<String> urls = ListUtil.fromArray(StringUtil.splitLines(content));
-
-		List<String> outURLs = new ArrayList<>();
-
-		for (String url : urls) {
-			outURLs.add(url);
-
-			if (StringUtil.indexOfAny(url, findParams) > -1) {
-				outURLs.add(StringUtil.replace(url, findParams, addParams));
-			}
-		}
-
-		return StringUtil.merge(outURLs, StringPool.NEW_LINE);
-	}
-
 	private void _exportImportLayouts(boolean privateLayout) throws Exception {
 		List<Layout> layouts = LayoutLocalServiceUtil.getLayouts(
 			_stagingGroup.getGroupId(), privateLayout);
@@ -1075,67 +1108,6 @@ public class DefaultExportImportContentProcessorTest {
 
 		ExportImportLocalServiceUtil.importLayouts(
 			exportImportConfiguration, larFile);
-	}
-
-	private String _extractValidContent(String content) {
-		return StringUtil.merge(
-			TransformUtil.transform(
-				ListUtil.fromArray(StringUtil.splitLines(content)),
-				line -> {
-					if (Validator.isNotNull(line) &&
-						!line.endsWith(StringPool.COLON)) {
-
-						return line;
-					}
-
-					return null;
-				}),
-			StringPool.NEW_LINE);
-	}
-
-	private String _getContent(String fileName) throws Exception {
-		Class<?> clazz = getClass();
-
-		InputStream inputStream = clazz.getResourceAsStream(
-			"dependencies/" + fileName);
-
-		Scanner scanner = new Scanner(inputStream);
-
-		scanner.useDelimiter("\\Z");
-
-		return scanner.next();
-	}
-
-	private Locale _getNondefaultLocale() throws Exception {
-		for (Locale locale : _locales) {
-			if (!locale.equals(_defaultLocale)) {
-				return locale;
-			}
-		}
-
-		throw new Exception("Could not find a non-default locale");
-	}
-
-	private List<String> _getURLs(String content) {
-		Matcher matcher = _pattern.matcher(StringPool.BLANK);
-
-		return TransformUtil.transformToList(
-			StringUtil.split(content, StringPool.NEW_LINE),
-			line -> {
-				matcher.reset(line);
-
-				if (matcher.find()) {
-					return line;
-				}
-
-				return null;
-			});
-	}
-
-	private String _replaceExternalGroupFriendlyURLs(String content) {
-		return _duplicateLinesWithParamNames(
-			content, _GROUP_FRIENDLY_URL_VARIABLES,
-			_EXTERNAL_GROUP_FRIENDLY_URL_VARIABLES);
 	}
 
 	private String _replaceLinksToLayoutsParameters(
@@ -1174,155 +1146,6 @@ public class DefaultExportImportContentProcessorTest {
 			});
 	}
 
-	private String _replaceMultiLocaleLayoutFriendlyURLs(String content) {
-		return _duplicateLinesWithParamNames(
-			content, _MULTI_LOCALE_LAYOUT_VARIABLES,
-			_NONDEFAULT_MULTI_LOCALE_LAYOUT_VARIABLES);
-	}
-
-	private String _replaceParameters(String content, FileEntry fileEntry) {
-		Company company = CompanyLocalServiceUtil.fetchCompany(
-			fileEntry.getCompanyId());
-
-		content = _replaceExternalGroupFriendlyURLs(content);
-		content = _replaceMultiLocaleLayoutFriendlyURLs(content);
-
-		Map<Locale, String> livePublicLayoutFriendlyURLMap =
-			_livePublicLayout.getFriendlyURLMap();
-		Map<Locale, String> stagingPrivateLayoutFriendlyURLMap =
-			_stagingPrivateLayout.getFriendlyURLMap();
-
-		LayoutSet stagingPrivateLayoutSet = _stagingGroup.getPrivateLayoutSet();
-
-		NavigableMap<String, String> stagingPrivateVirtualHostnames =
-			stagingPrivateLayoutSet.getVirtualHostnames();
-
-		Map<Locale, String> stagingPublicLayoutFriendlyURLMap =
-			_stagingPublicLayout.getFriendlyURLMap();
-
-		LayoutSet stagingPublicLayoutSet = _stagingGroup.getPublicLayoutSet();
-
-		NavigableMap<String, String> stagingPublicVirtualHostnames =
-			stagingPublicLayoutSet.getVirtualHostnames();
-
-		content = StringUtil.replace(
-			content,
-			new String[] {
-				"[$BLOG_ENTRY_DISPLAY_SERVLET_MAPPING$]",
-				"[$CANONICAL_URL_SEPARATOR$]", "[$CONTROL_PANEL_FRIENDLY_URL$]",
-				"[$CONTROL_PANEL_LAYOUT_FRIENDLY_URL$]",
-				"[$DL_ENTRY_DISPLAY_SERVLET_MAPPING$]",
-				"[$EXTERNAL_GROUP_FRIENDLY_URL$]",
-				"[$EXTERNAL_PRIVATE_LAYOUT_FRIENDLY_URL$]",
-				"[$EXTERNAL_PUBLIC_LAYOUT_FRIENDLY_URL$]",
-				"[$FILE_ENTRY_FRIENDLY_URL$]", "[$FILE_NAME$]",
-				"[$FRIENDLY_URL_SEPARATOR$]", "[$GROUP_FRIENDLY_URL$]",
-				"[$GROUP_ID$]", "[$GROUP_NAME$]",
-				"[$GROUP_PRIVATE_PAGES_VIRTUAL_HOST$]",
-				"[$GROUP_PUBLIC_PAGES_VIRTUAL_HOST$]", "[$IMAGE_ID$]",
-				"[$LIVE_GROUP_FRIENDLY_URL$]", "[$LIVE_GROUP_ID$]",
-				"[$LIVE_PUBLIC_LAYOUT_FRIENDLY_URL$]",
-				"[$NON_DEFAULT_LIVE_PUBLIC_LAYOUT_FRIENDLY_URL$]",
-				"[$NON_DEFAULT_PRIVATE_LAYOUT_FRIENDLY_URL$]",
-				"[$NON_DEFAULT_PUBLIC_LAYOUT_FRIENDLY_URL$]",
-				"[$NONREPLACEABLE_PRIVATE_LAYOUT_FRIENDLY_URL$]",
-				"[$NONREPLACEABLE_PUBLIC_LAYOUT_FRIENDLY_URL$]",
-				"[$PATH_CONTEXT$]", "[$PATH_FRIENDLY_URL_PRIVATE_GROUP$]",
-				"[$PATH_FRIENDLY_URL_PRIVATE_USER$]",
-				"[$PATH_FRIENDLY_URL_PUBLIC$]",
-				"[$PRIVATE_LAYOUT_FRIENDLY_URL$]",
-				"[$PUBLIC_LAYOUT_FRIENDLY_URL$]", "[$TITLE$]", "[$UUID$]",
-				"[$WEB_CONTENT_DISPLAY_SERVLET_MAPPING$]", "[$WEB_ID$]"
-			},
-			new String[] {
-				FriendlyURLResolverConstants.URL_SEPARATOR_X_BLOGS_ENTRY,
-				VirtualLayoutConstants.CANONICAL_URL_SEPARATOR,
-				GroupConstants.CONTROL_PANEL_FRIENDLY_URL,
-				PropsValues.CONTROL_PANEL_LAYOUT_FRIENDLY_URL,
-				FriendlyURLResolverConstants.URL_SEPARATOR_X_FILE_ENTRY,
-				_externalGroup.getFriendlyURL(),
-				_externalPrivateLayout.getFriendlyURL(),
-				_externalPublicLayout.getFriendlyURL(),
-				FriendlyURLNormalizerUtil.normalizeWithPeriodsAndSlashes(
-					fileEntry.getTitle()),
-				fileEntry.getFileName(), Portal.FRIENDLY_URL_SEPARATOR,
-				_stagingGroup.getFriendlyURL(),
-				String.valueOf(fileEntry.getGroupId()),
-				StringUtil.removeFirst(
-					_stagingGroup.getFriendlyURL(), StringPool.SLASH),
-				stagingPrivateVirtualHostnames.firstKey(),
-				stagingPublicVirtualHostnames.firstKey(),
-				String.valueOf(fileEntry.getFileEntryId()),
-				_liveGroup.getFriendlyURL(),
-				String.valueOf(_liveGroup.getGroupId()),
-				_livePublicLayout.getFriendlyURL(),
-				livePublicLayoutFriendlyURLMap.get(_nondefaultLocale),
-				stagingPrivateLayoutFriendlyURLMap.get(_nondefaultLocale),
-				stagingPublicLayoutFriendlyURLMap.get(_nondefaultLocale),
-				_stagingPrivateLayout.getFriendlyURL(),
-				_stagingPublicLayout.getFriendlyURL(),
-				PortalUtil.getPathContext(),
-				PropsValues.LAYOUT_FRIENDLY_URL_PRIVATE_GROUP_SERVLET_MAPPING,
-				PropsValues.LAYOUT_FRIENDLY_URL_PRIVATE_USER_SERVLET_MAPPING,
-				PropsValues.LAYOUT_FRIENDLY_URL_PUBLIC_SERVLET_MAPPING,
-				_stagingPrivateLayout.getFriendlyURL(),
-				_stagingPublicLayout.getFriendlyURL(), fileEntry.getTitle(),
-				fileEntry.getUuid(),
-				FriendlyURLResolverConstants.URL_SEPARATOR_X_JOURNAL_ARTICLE,
-				company.getWebId()
-			});
-
-		if (!content.contains("[$TIMESTAMP")) {
-			return _extractValidContent(content);
-		}
-
-		return _replaceTimestampParameters(content);
-	}
-
-	private String _replaceTimestampParameters(String content) {
-		List<String> urls = ListUtil.fromArray(StringUtil.splitLines(content));
-
-		String timestampParameter = "t=123456789";
-
-		String parameters1 = timestampParameter + "&width=100&height=100";
-		String parameters2 = "width=100&" + timestampParameter + "&height=100";
-		String parameters3 = "width=100&height=100&" + timestampParameter;
-		String parameters4 = StringBundler.concat(
-			timestampParameter, "?", timestampParameter,
-			"&width=100&height=100");
-
-		List<String> outURLs = new ArrayList<>();
-
-		for (String url : urls) {
-			if (Validator.isNotNull(url) && !url.contains("[$TIMESTAMP") &&
-				!url.endsWith(StringPool.COLON)) {
-
-				outURLs.add(url);
-
-				continue;
-			}
-
-			outURLs.add(
-				StringUtil.replace(
-					url, new String[] {"[$TIMESTAMP$]", "[$TIMESTAMP_ONLY$]"},
-					new String[] {"&" + parameters1, "?" + parameters1}));
-			outURLs.add(
-				StringUtil.replace(
-					url, new String[] {"[$TIMESTAMP$]", "[$TIMESTAMP_ONLY$]"},
-					new String[] {"&" + parameters2, "?" + parameters2}));
-			outURLs.add(
-				StringUtil.replace(
-					url, new String[] {"[$TIMESTAMP$]", "[$TIMESTAMP_ONLY$]"},
-					new String[] {"&" + parameters3, "?" + parameters3}));
-			outURLs.add(
-				StringUtil.replace(
-					url, new String[] {"[$TIMESTAMP$]", "[$TIMESTAMP_ONLY$]"},
-					new String[] {StringPool.BLANK, "?" + parameters4}));
-		}
-
-		return StringUtil.merge(outURLs, StringPool.NEW_LINE);
-	}
-
 	private void _testImportDLReferences(boolean deleteFileEntryBeforeImport)
 		throws Exception {
 
@@ -1336,8 +1159,14 @@ public class DefaultExportImportContentProcessorTest {
 		referrerStagedModelElement.addAttribute(
 			"path", referrerStagedModelPath);
 
-		String content = _replaceParameters(
-			_getContent("dl_references.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "dl_references.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		content = _exportImportContentProcessor.replaceExportContentReferences(
 			_portletDataContextExport, _referrerStagedModel, content, true,
@@ -1378,8 +1207,14 @@ public class DefaultExportImportContentProcessorTest {
 		referrerStagedModelElement.addAttribute(
 			"path", referrerStagedModelPath);
 
-		String content = _replaceParameters(
-			_getContent("dl_references_file_friendly_urls.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "dl_references_file_friendly_urls.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		content = _exportImportContentProcessor.replaceExportContentReferences(
 			_portletDataContextExport, _referrerStagedModel, content, true,
@@ -1398,8 +1233,14 @@ public class DefaultExportImportContentProcessorTest {
 	}
 
 	private void _testImportLayoutReferences() throws Exception {
-		String content = _replaceParameters(
-			_getContent("layout_references.txt"), _fileEntry);
+		String content = ExportImportContentProcessorTestUtil.replaceParameters(
+			_externalGroup,
+			new ObjectValuePair<>(
+				_externalPrivateLayout, _externalPublicLayout),
+			_fileEntry, "layout_references.txt", _stagingGroup,
+			new ObjectValuePair<>(_stagingPrivateLayout, _stagingPublicLayout),
+			_liveGroup,
+			new ObjectValuePair<>(_livePrivateLayout, _livePublicLayout));
 
 		_exportImportContentProcessor.validateContentReferences(
 			_stagingGroup.getGroupId(), content);
@@ -1426,17 +1267,6 @@ public class DefaultExportImportContentProcessorTest {
 			content, content.contains("data_handler_site_admin_url"));
 	}
 
-	private static final String[] _EXTERNAL_GROUP_FRIENDLY_URL_VARIABLES = {
-		"[$EXTERNAL_GROUP_FRIENDLY_URL$]",
-		"[$EXTERNAL_PRIVATE_LAYOUT_FRIENDLY_URL$]",
-		"[$EXTERNAL_PUBLIC_LAYOUT_FRIENDLY_URL$]"
-	};
-
-	private static final String[] _GROUP_FRIENDLY_URL_VARIABLES = {
-		"[$GROUP_FRIENDLY_URL$]", "[$PRIVATE_LAYOUT_FRIENDLY_URL$]",
-		"[$PUBLIC_LAYOUT_FRIENDLY_URL$]"
-	};
-
 	private static final char[] _LAYOUT_REFERENCE_STOP_CHARS = {
 		CharPool.APOSTROPHE, CharPool.CLOSE_BRACKET, CharPool.CLOSE_CURLY_BRACE,
 		CharPool.CLOSE_PARENTHESIS, CharPool.GREATER_THAN, CharPool.LESS_THAN,
@@ -1444,23 +1274,7 @@ public class DefaultExportImportContentProcessorTest {
 		CharPool.SPACE
 	};
 
-	private static final String[] _MULTI_LOCALE_LAYOUT_VARIABLES = {
-		"[$LIVE_PUBLIC_LAYOUT_FRIENDLY_URL$]",
-		"[$PRIVATE_LAYOUT_FRIENDLY_URL$]", "[$PUBLIC_LAYOUT_FRIENDLY_URL$]"
-	};
-
-	private static final String[] _NONDEFAULT_MULTI_LOCALE_LAYOUT_VARIABLES = {
-		"[$NON_DEFAULT_LIVE_PUBLIC_LAYOUT_FRIENDLY_URL$]",
-		"[$NON_DEFAULT_PRIVATE_LAYOUT_FRIENDLY_URL$]",
-		"[$NON_DEFAULT_PUBLIC_LAYOUT_FRIENDLY_URL$]"
-	};
-
-	private static final Locale[] _locales = {
-		LocaleUtil.US, LocaleUtil.GERMANY, LocaleUtil.SPAIN
-	};
 	private static String _oldLayoutFriendlyURLPrivateUserServletMapping;
-	private static final Pattern _pattern = Pattern.compile(
-		"href=|url\\(|\\{|\\[");
 
 	private Locale _defaultLocale;
 	private ExportImportContentProcessor<String> _exportImportContentProcessor;
@@ -1481,7 +1295,6 @@ public class DefaultExportImportContentProcessorTest {
 
 	private Layout _livePrivateLayout;
 	private Layout _livePublicLayout;
-	private Locale _nondefaultLocale;
 	private PortletDataContext _portletDataContextExport;
 	private PortletDataContext _portletDataContextImport;
 	private StagedModel _referrerStagedModel;
